Moved jb_utils to renamed notebook_utils
