+++ openmdao/recorders/sqlite_reader.py
@@ -16,7 +16,7 @@ from openmdao.utils.record_util import check_valid_sqlite3_db, get_source_system
 
 from openmdao.recorders.sqlite_recorder import format_version, META_KEY_SEP
 
-from openmdao.core.notebook_mode import notebook, tabulate
+from openmdao.core.notebook_utils import notebook, tabulate
 try:
     from IPython.display import display, HTML
 except ImportError:
